Add support for closing invoice date to tenant
This change adds support for the closing invoice date to a tenant.
The closing invoice date is a new attribute in the DynamoDB table. The
TenantConfiguration class now supports retrieiving it and handling it
if it isn't there.

This commit also updates tests that depend on the TenantConfiguration
class. It adds tests to address the additional TenantConfiguration
functionality.